Avoid cross-thread timing issues
This adds a prompt and avoids issues where reading and writing are
dependent on one another. Previously, if the device wouldn't send any
output, the output handler would just sit on its mutex, blocking the
input handler. Instead drop the massive timeout, timeout early and
release the mutex again if there is no data.

This is slightly hacky because there's a manual waiting time on the
input handler to avoid printing the prompt while there's still output
from the last command printing. A neater solution would be to also
wrap stdout into a mutex and just have it wait, but this is good
enough for now.